core: network-manager: Spread out implementation for discoverability
