+++ core/src/network_manager/manager.rs
@@ -1,24 +1,22 @@
 //! The network manager handles lower level interaction with the p2p network
+mod control_directives;
+mod identify;
+mod pubsub;
+mod request_response;
 
 use ed25519_dalek::Keypair as SigKeypair;
 use futures::StreamExt;
-use itertools::Itertools;
 use libp2p::{
-    gossipsub::{Event as GossipsubEvent, Message as GossipsubMessage, Sha256Topic},
-    identify::Event as IdentifyEvent,
+    gossipsub::{Event as GossipsubEvent, Sha256Topic},
     identity::Keypair,
     multiaddr::Protocol,
-    request_response::{Event as RequestResponseEvent, Message as RequestResponseMessage},
+    request_response::Event as RequestResponseEvent,
     swarm::SwarmEvent,
-    Multiaddr, PeerId, Swarm,
+    Multiaddr, Swarm,
 };
-use libp2p_core::Endpoint;
-use libp2p_swarm::{ConnectionId, NetworkBehaviour};
-use mpc_ristretto::network::{MpcNetwork, QuicTwoPartyNet};
 use portpicker::Port;
 use tokio::sync::mpsc::UnboundedSender as TokioSender;
 use tracing::log;
-use uuid::Uuid;
 
 use std::{net::SocketAddr, thread::JoinHandle};
 use tokio::sync::mpsc::UnboundedReceiver;
@@ -26,17 +24,15 @@ use tokio::sync::mpsc::UnboundedReceiver;
 use crate::{
     default_wrapper::DefaultWrapper,
     gossip::{
-        jobs::{ClusterManagementJob, GossipServerJob, OrderBookManagementJob},
+        jobs::GossipServerJob,
         types::{ClusterId, PeerInfo, WrappedPeerId},
     },
     gossip_api::{
-        cluster_management::{ClusterManagementMessage, ReplicatedMessage},
         gossip::{
-            AuthenticatedGossipRequest, AuthenticatedGossipResponse, AuthenticatedPubsubMessage,
-            ConnectionRole, GossipOutbound, GossipOutbound::Pubsub, GossipRequest, GossipResponse,
-            ManagerControlDirective, PubsubMessage,
+            AuthenticatedGossipRequest, AuthenticatedGossipResponse, GossipOutbound,
+            GossipOutbound::Pubsub, PubsubMessage,
         },
-        orderbook_management::{OrderBookManagementMessage, OrderInfoResponse, ORDER_BOOK_TOPIC},
+        orderbook_management::ORDER_BOOK_TOPIC,
     },
     handshake::jobs::HandshakeExecutionJob,
     state::RelayerState,
@@ -377,538 +373,6 @@ impl NetworkManagerExecutor {
             Pubsub { topic, message } => self.forward_outbound_pubsub(topic, message),
             GossipOutbound::ManagementMessage(command) => self.handle_control_directive(command),
         }
-    }
-
-    /// Forward an outbound pubsub message to the network
-    fn forward_outbound_pubsub(
-        &mut self,
-        topic: String,
-        message: PubsubMessage,
-    ) -> Result<(), NetworkManagerError> {
-        // If the gossip server has not warmed up the local node into the network, buffer
-        // the pubsub message for forwarding after the warmup
-        if !self.warmup_finished {
-            self.warmup_buffer
-                .push(BufferedPubsubMessage { topic, message });
-            return Ok(());
-        }
-
-        // If we require a signature on the message attach one
-        let req_body = AuthenticatedPubsubMessage::new_with_body(message, &self.cluster_key)
-            .map_err(|err| NetworkManagerError::Authentication(err.to_string()))?;
-
-        // Forward to the network
-        let topic = Sha256Topic::new(topic);
-        self.swarm
-            .behaviour_mut()
-            .pubsub
-            .publish(topic, req_body)
-            .map_err(|err| NetworkManagerError::Network(err.to_string()))?;
-        Ok(())
-    }
-
-    // ------------------------------
-    // | Control Directive Handlers |
-    // ------------------------------
-
-    /// Handles a message from another worker module that explicitly directs the network manager
-    /// to take some action
-    ///
-    /// The end destination of these messages is not a network peer, but the local network manager
-    /// itself
-    fn handle_control_directive(
-        &mut self,
-        command: ManagerControlDirective,
-    ) -> Result<(), NetworkManagerError> {
-        match command {
-            // Register a new peer in the distributed routing tables
-            ManagerControlDirective::NewAddr { peer_id, address } => {
-                // If we cannot parse the address or it is not dialable, skip indexing
-                if !is_dialable_multiaddr(&address, self.allow_local) {
-                    log::info!("skipping local addr {:?}", address);
-                    return Ok(());
-                }
-
-                self.swarm
-                    .behaviour_mut()
-                    .kademlia_dht
-                    .add_address(&peer_id, address);
-
-                Ok(())
-            }
-
-            // Build an MPC net for the given peers to communicate over
-            ManagerControlDirective::BrokerMpcNet {
-                request_id,
-                peer_id,
-                peer_port,
-                local_port,
-                local_role,
-            } => {
-                // Lookup known remote addresses for the peer
-                let known_peer_addrs = self
-                    .swarm
-                    .behaviour_mut()
-                    .handle_pending_outbound_connection(
-                        ConnectionId::new_unchecked(0),
-                        Some(peer_id.0),
-                        &[],
-                        Endpoint::Dialer,
-                    )
-                    .map_err(|_| NetworkManagerError::Network(ERR_NO_KNOWN_ADDR.to_string()))?;
-
-                // Spawn a separate task to asynchronously dial the peer and respond to
-                // the handshake manager's brokerage request
-                let allow_local = self.allow_local;
-                let handshake_queue_clone = self.handshake_work_queue.clone();
-                tokio::spawn(async move {
-                    if let Err(e) = Self::broker_mpc_net(
-                        allow_local,
-                        request_id,
-                        peer_port,
-                        local_port,
-                        local_role,
-                        known_peer_addrs,
-                        handshake_queue_clone,
-                    )
-                    .await
-                    {
-                        log::error!("error brokering MPC network: {e}");
-                    }
-                });
-
-                Ok(())
-            }
-
-            // Inform the network manager that the gossip server has warmed up the local node in
-            // the cluster by advertising the local node's presence
-            //
-            // The network manager delays sending pubsub events until the gossip protocol has warmed
-            // up, because at startup, there are no known peers to publish to. The network manager gives
-            // the gossip server some time to discover new addresses before publishing to the network.
-            ManagerControlDirective::GossipWarmupComplete => {
-                self.warmup_finished = true;
-                // Forward all buffered messages to the network
-                for buffered_message in self.warmup_buffer.drain(..).collect_vec() {
-                    self.forward_outbound_pubsub(buffered_message.topic, buffered_message.message)?;
-                }
-
-                Ok(())
-            }
-        }
-    }
-
-    /// Broker an MPC net between the local and remote peer using the given known addresses
-    async fn broker_mpc_net(
-        allow_local: bool,
-        request_id: Uuid,
-        peer_port: u16,
-        mut local_port: u16,
-        local_role: ConnectionRole,
-        known_peer_addrs: Vec<Multiaddr>,
-        handshake_work_queue: TokioSender<HandshakeExecutionJob>,
-    ) -> Result<(), NetworkManagerError> {
-        // Connect on a side-channel to the peer
-        let party_id = local_role.get_party_id();
-
-        let mpc_net = match local_role {
-            ConnectionRole::Dialer => {
-                // Attempt to dial the peer on all known addresses
-                let mut brokered_net = None;
-                for peer_addr in known_peer_addrs
-                    .into_iter()
-                    .filter_map(|addr| multiaddr_to_socketaddr(&addr, peer_port))
-                    .filter(|addr| is_dialable_addr(addr, allow_local))
-                {
-                    let local_addr: SocketAddr =
-                        format!("127.0.0.1:{:?}", local_port).parse().unwrap();
-                    let mut net = QuicTwoPartyNet::new(party_id, local_addr, peer_addr);
-
-                    if let Err(e) = net.connect().await {
-                        log::warn!(
-                            "failed to broker MPC network on address {peer_addr:?}, error: {e}"
-                        )
-                    } else {
-                        log::info!("successfully connected to peer at addr: {peer_addr:?}");
-                        // Send a wakeup over the brokered bi-stream, this is not automatically done
-                        net.send_bytes(&[0u8])
-                            .await
-                            .map_err(|err| NetworkManagerError::Network(err.to_string()))?;
-
-                        // Forward the net to the handshake manager
-                        brokered_net = Some(net);
-                        break;
-                    }
-
-                    // During the connection attempt, the selected port was bound to, and dropping the `net`
-                    // may not immediately free up the port; so we increment the outbound port and continue
-                    local_port += 1;
-                }
-
-                brokered_net
-                    .ok_or_else(|| NetworkManagerError::Network(ERR_BROKER_MPC_NET.to_string()))?
-            }
-
-            ConnectionRole::Listener => {
-                // As the listener, the peer address is inconsequential, and can be a dummy value
-                let local_addr: SocketAddr = format!("0.0.0.0:{local_port}").parse().unwrap();
-                let peer_addr: SocketAddr = "0.0.0.0:0".parse().unwrap();
-                let mut net = QuicTwoPartyNet::new(party_id, local_addr, peer_addr);
-                net.connect()
-                    .await
-                    .map_err(|err| NetworkManagerError::Network(err.to_string()))?;
-
-                // Clear the wakeup buffer
-                net.receive_bytes()
-                    .await
-                    .map_err(|err| NetworkManagerError::Network(err.to_string()))?;
-                net
-            }
-        };
-
-        // After the dependencies are injected into the network; forward it to the handshake manager to
-        // dial the peer and begin the MPC
-        handshake_work_queue
-            .send(HandshakeExecutionJob::MpcNetSetup {
-                request_id,
-                party_id,
-                net: mpc_net,
-            })
-            .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string()))?;
-        Ok(())
-    }
-
-    // ------------------------------
-    // | Identify Protocol Handlers |
-    // ------------------------------
-
-    /// Handle a message from the Identify protocol
-    async fn handle_identify_event(
-        &mut self,
-        event: IdentifyEvent,
-    ) -> Result<(), NetworkManagerError> {
-        // Update the local peer's public IP address if it has not already been discovered
-        if let IdentifyEvent::Received { info, .. } = event {
-            if !self.discovered_identity {
-                // Replace the port if the discovered NAT port is incorrect
-                let mut local_addr = info.observed_addr;
-                replace_port(&mut local_addr, self.p2p_port);
-
-                // Add the p2p multihash to the multiaddr
-                local_addr = local_addr.with(Protocol::P2p(self.local_peer_id.0.into()));
-
-                // Update the addr in the global state
-                log::info!("discovered local peer's public IP: {:?}", local_addr);
-                self.global_state.update_local_peer_addr(local_addr).await;
-                self.discovered_identity = true;
-
-                // Optimistically broadcast the discovered identity to the network via
-                // the heartbeat sub-protocol
-                for peer in self.global_state.read_peer_index().await.get_all_peer_ids() {
-                    if let Err(e) = self
-                        .gossip_work_queue
-                        .send(GossipServerJob::ExecuteHeartbeat(peer))
-                    {
-                        log::error!("error forwarding heartbeat to gossip server: {e}")
-                    }
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    // -----------------------------
-    // | Request/Response Handlers |
-    // -----------------------------
-
-    /// Handle an incoming message from the network's request/response protocol
-    fn handle_inbound_request_response_message(
-        &mut self,
-        peer_id: PeerId,
-        message: RequestResponseMessage<AuthenticatedGossipRequest, AuthenticatedGossipResponse>,
-    ) -> Result<(), NetworkManagerError> {
-        // Multiplex over request/response message types
-        match message {
-            // Handle inbound request from another peer
-            RequestResponseMessage::Request {
-                request, channel, ..
-            } => {
-                // Authenticate the request
-                if !request.verify_cluster_auth(&self.cluster_key.public) {
-                    return Err(NetworkManagerError::Authentication(
-                        ERR_SIG_VERIFY.to_string(),
-                    ));
-                }
-
-                match request.body {
-                    // Forward the bootstrap request directly to the gossip server
-                    GossipRequest::Bootstrap(req) => self
-                        .gossip_work_queue
-                        .send(GossipServerJob::Bootstrap(req, channel))
-                        .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string())),
-
-                    GossipRequest::Heartbeat(heartbeat_message) => self
-                        .gossip_work_queue
-                        .send(GossipServerJob::HandleHeartbeatReq {
-                            peer_id: WrappedPeerId(peer_id),
-                            message: heartbeat_message,
-                            channel,
-                        })
-                        .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string())),
-
-                    GossipRequest::Handshake {
-                        request_id,
-                        message,
-                    } => self
-                        .handshake_work_queue
-                        .send(HandshakeExecutionJob::ProcessHandshakeMessage {
-                            request_id,
-                            peer_id: WrappedPeerId(peer_id),
-                            message,
-                            response_channel: Some(channel),
-                        })
-                        .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string())),
-
-                    GossipRequest::OrderInfo(req) => self
-                        .gossip_work_queue
-                        .send(GossipServerJob::OrderBookManagement(
-                            OrderBookManagementJob::OrderInfo {
-                                order_id: req.order_id,
-                                response_channel: channel,
-                            },
-                        ))
-                        .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string())),
-
-                    GossipRequest::Replicate(replicate_message) => {
-                        self.gossip_work_queue
-                            .send(GossipServerJob::Cluster(
-                                ClusterManagementJob::ReplicateRequest(replicate_message),
-                            ))
-                            .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string()))?;
-
-                        // Send a simple ack back to avoid closing the channel
-                        self.swarm
-                            .behaviour_mut()
-                            .request_response
-                            .send_response(channel, AuthenticatedGossipResponse::new_ack())
-                            .map_err(|_| {
-                                NetworkManagerError::Network("error sending Ack".to_string())
-                            })
-                    }
-
-                    GossipRequest::ValidityProof {
-                        order_id,
-                        proof_bundle,
-                    } => self
-                        .gossip_work_queue
-                        .send(GossipServerJob::Cluster(
-                            ClusterManagementJob::UpdateValidityProof(order_id, proof_bundle),
-                        ))
-                        .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string())),
-
-                    GossipRequest::ValidityWitness { order_id, witness } => {
-                        self.gossip_work_queue
-                            .send(GossipServerJob::OrderBookManagement(
-                                OrderBookManagementJob::OrderWitnessResponse { order_id, witness },
-                            ))
-                            .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string()))?;
-
-                        // Send back an ack
-                        self.handle_outbound_message(GossipOutbound::Response {
-                            channel,
-                            message: GossipResponse::Ack,
-                        })
-                    }
-
-                    GossipRequest::WalletUpdate { wallet } => {
-                        self.gossip_work_queue
-                            .send(GossipServerJob::WalletUpdate { wallet })
-                            .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string()))?;
-
-                        // Send back an ack
-                        self.handle_outbound_message(GossipOutbound::Response {
-                            channel,
-                            message: GossipResponse::Ack,
-                        })
-                    }
-                }
-            }
-
-            // Handle inbound response
-            RequestResponseMessage::Response { response, .. } => {
-                if !response.verify_cluster_auth(&self.cluster_key.public) {
-                    return Err(NetworkManagerError::Authentication(
-                        ERR_SIG_VERIFY.to_string(),
-                    ));
-                }
-
-                match response.body {
-                    GossipResponse::Ack => Ok(()),
-
-                    GossipResponse::Heartbeat(heartbeat_message) => self
-                        .gossip_work_queue
-                        .send(GossipServerJob::HandleHeartbeatResp {
-                            peer_id: WrappedPeerId(peer_id),
-                            message: heartbeat_message,
-                        })
-                        .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string())),
-
-                    GossipResponse::Handshake {
-                        request_id,
-                        message,
-                    } => self
-                        .handshake_work_queue
-                        .send(HandshakeExecutionJob::ProcessHandshakeMessage {
-                            request_id,
-                            peer_id: WrappedPeerId(peer_id),
-                            message,
-                            // The handshake should response via a new request sent on the network manager channel
-                            response_channel: None,
-                        })
-                        .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string())),
-
-                    GossipResponse::OrderInfo(OrderInfoResponse { order_id, info }) => self
-                        .gossip_work_queue
-                        .send(GossipServerJob::OrderBookManagement(
-                            OrderBookManagementJob::OrderInfoResponse { order_id, info },
-                        ))
-                        .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string())),
-                }
-            }
-        }
-    }
-
-    // ------------------
-    // | Pubsub Handlers |
-    // -------------------
-
-    /// Handle an incoming network request for a pubsub message
-    fn handle_inbound_pubsub_message(
-        &mut self,
-        message: GossipsubMessage,
-    ) -> Result<(), NetworkManagerError> {
-        // Deserialize into API types and verify auth
-        let event: AuthenticatedPubsubMessage = message.data.into();
-        if !event.verify_cluster_auth(&self.cluster_key.public) {
-            return Err(NetworkManagerError::Authentication(
-                ERR_SIG_VERIFY.to_string(),
-            ));
-        }
-
-        match event.body {
-            PubsubMessage::ClusterManagement {
-                cluster_id,
-                message,
-            } => {
-                match message {
-                    // --------------------
-                    // | Cluster Metadata |
-                    // --------------------
-
-                    // Forward the management message to the gossip server for processing
-                    ClusterManagementMessage::Join(join_request) => {
-                        // Forward directly
-                        self.gossip_work_queue
-                            .send(GossipServerJob::Cluster(
-                                ClusterManagementJob::ClusterJoinRequest(cluster_id, join_request),
-                            ))
-                            .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string()))?;
-                    }
-
-                    // Forward the management message to the gossip server for processing
-                    ClusterManagementMessage::Replicated(ReplicatedMessage {
-                        wallets,
-                        peer_id,
-                    }) => {
-                        // Forward one job per replicated wallet; makes gossip server implementation cleaner
-                        for wallet_id in wallets.into_iter() {
-                            self.gossip_work_queue
-                                .send(GossipServerJob::Cluster(
-                                    ClusterManagementJob::AddWalletReplica { wallet_id, peer_id },
-                                ))
-                                .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string()))?;
-                        }
-                    }
-
-                    // ---------
-                    // | Match |
-                    // ---------
-
-                    // Forward the cache sync message to the handshake manager to update the local
-                    // cache copy
-                    ClusterManagementMessage::CacheSync(order1, order2) => self
-                        .handshake_work_queue
-                        .send(HandshakeExecutionJob::CacheEntry { order1, order2 })
-                        .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string()))?,
-
-                    // Forward the match in progress message to the handshake manager so that it can avoid
-                    // scheduling a duplicate handshake for the given order pair
-                    ClusterManagementMessage::MatchInProgress(order1, order2) => self
-                        .handshake_work_queue
-                        .send(HandshakeExecutionJob::PeerMatchInProgress { order1, order2 })
-                        .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string()))?,
-
-                    // -------------
-                    // | Orderbook |
-                    // -------------
-
-                    // Forward a request for validity proofs to the gossip server to check for locally
-                    // available proofs
-                    ClusterManagementMessage::RequestOrderValidityProof(req) => {
-                        self.gossip_work_queue
-                            .send(GossipServerJob::Cluster(
-                                ClusterManagementJob::ShareValidityProofs(req),
-                            ))
-                            .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string()))?;
-                    }
-
-                    //Forward a request to the gossip server to share validity proof witness
-                    ClusterManagementMessage::RequestOrderValidityWitness(req) => self
-                        .gossip_work_queue
-                        .send(GossipServerJob::OrderBookManagement(
-                            OrderBookManagementJob::OrderWitness {
-                                order_id: req.order_id,
-                                requesting_peer: req.sender,
-                            },
-                        ))
-                        .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string()))?,
-                }
-            }
-            PubsubMessage::OrderBookManagement(msg) => match msg {
-                OrderBookManagementMessage::OrderReceived {
-                    order_id,
-                    nullifier,
-                    cluster,
-                } => self
-                    .gossip_work_queue
-                    .send(GossipServerJob::OrderBookManagement(
-                        OrderBookManagementJob::OrderReceived {
-                            order_id,
-                            nullifier,
-                            cluster,
-                        },
-                    ))
-                    .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string()))?,
-
-                OrderBookManagementMessage::OrderProofUpdated {
-                    order_id,
-                    cluster,
-                    proof_bundle,
-                } => self
-                    .gossip_work_queue
-                    .send(GossipServerJob::OrderBookManagement(
-                        OrderBookManagementJob::OrderProofUpdated {
-                            order_id,
-                            cluster,
-                            proof_bundle,
-                        },
-                    ))
-                    .map_err(|err| NetworkManagerError::EnqueueJob(err.to_string()))?,
-            },
-        }
-
-        Ok(())
     }
 }
